requirements.txt
langchain==0.0.332
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.4.0
protobuf~=3.19.0
chromadb==0.4.15
InstructorEmbedding
langchain-google-genai
google-generativeai
sentence-transformers==2.2.2
langchain_community